Got user and notes backend connected to front end. Got token settings working. Set up private routes.

+++ routes/notes.js
@@ -1,62 +1,88 @@
 const express = require('express');
 const router = express.Router();
 const client = require('../database')
+const auth = require('../middleware/authentication')
 
-router.get('/', (req, res) => {
-    client.query('SELECT * FROM notes', (error, results) => {
+router.get('/', auth, (req, res) => {
+    client.query('SELECT * FROM notes WHERE author=$1', [req.user.username], (error, results) => {
         if (error){
-            console.log(error);
+            return res.status(400).json({msg: "Could not get notes."})
         }
         else{
-            res.status(200).json(results.rows)
+            return res.status(200).json(results.rows)
         }
     })
 })
 
-router.get('/:id', (req, res) => {
-    client.query('SELECT * FROM notes WHERE note_id=$1', [req.params.id], (error, results) => {
+router.get('/:id', auth, (req, res) => {
+    client.query('SELECT * FROM notes WHERE note_id=$1 AND author=$2', [req.params.id, req.user.username], (error, results) => {
         if (error){
-            console.log(error);
+            return res.status(400).json({msg: "Could not get note."})
         }
         else{
-            res.status(200).json(results.rows)
+            return res.status(200).json(results.rows[0])
         }
     })
 })
 
-router.post('/', (req, res) => {
-    values = [req.body.date, req.body.title, req.body.body]
-    client.query('INSERT INTO notes(date, title, body) VALUES($1, $2, $3)', values, (error, results) => {
+router.post('/', auth, (req, res) => {
+    values = [req.body.date, req.body.title, req.body.body, req.body.author]
+    client.query('INSERT INTO notes(date, title, body, author) VALUES($1, $2, $3, $4) RETURNING *', values, (error, results) => {
         if (error){
-            console.log(error);
+            return res.status(400).json({msg: "Could not create note."})
         }
         else{
-            res.status(201).json({msg: "Entry has been created"})
+            return res.status(201).json({msg: "Note has been created.", results})
         }
     })
 })
 
-router.delete('/:id', (req, res) => {
-    client.query('DELETE FROM notes WHERE note_id=$1', [req.params.id], (error, results) => {
+router.delete('/:id', auth, (req, res) => {
+    console.log("here")
+    client.query('DELETE FROM notes WHERE (note_id=$1 AND author=$2)', [req.params.id, req.user.username], (error, results) => {
         if (error){
-            console.log(error);
+            return res.status(400).json({msg: "Could not delete note."})
         }
         else{
-            res.status(202).json({msg: "Entry has been deleted."})
+            return res.status(202).json({msg: "Entry has been deleted."})
         }
     })
 })
 
-router.put('/:id', (req, res) => {
-    values = [req.body.title, req.body.body, req.params.id]
-    client.query('UPDATE notes SET title=$1, body=$2 WHERE note_id=$3' , values, (error, results) => {
-        if (error){
-            console.log(error);
-        }
-        else{
-            res.status(200).json({msg: "Entry has been updated"})
-        }
-    })
+router.put('/:id', auth, (req, res) => {
+    if (req.body.title && req.body.body){
+        values = [req.body.title, req.body.body, req.user.username, req.params.id]
+        client.query('UPDATE notes SET title=$1, body=$2 WHERE (author=$3 AND note_id=$4)' , values, (error, results) => {
+            if (error){
+                return res.status(401).json({msg: "Could not update note."})
+            }
+            else{
+                return res.status(200).json({msg: "Note has been updated."})
+            }
+        })
+    }
+    else if (req.body.body){
+        values = [req.body.body, req.user.username, req.params.id]
+        client.query('UPDATE notes SET body=$1 WHERE (author=$2 AND note_id=$3)' , values, (error, results) => {
+            if (error){
+                return res.status(401).json({msg: "Could not update note."})
+            }
+            else{
+                return res.status(200).json({msg: "Note has been updated."})
+            }
+        })
+    }
+    else if (req.body.title){
+        values = [req.body.title, req.user.username, req.params.id]
+        client.query('UPDATE notes SET title=$1 WHERE (author=$2 AND note_id=$3)' , values, (error, results) => {
+            if (error){
+                return res.status(401).json({msg: "Could not update note."})
+            }
+            else{
+                return res.status(200).json({msg: "Note has been updated."})
+            }
+        })
+    }
 })
 
 module.exports = router
